convert to double

+++ FieldSolver/bfield.c
@@ -16,19 +16,19 @@ July 10, 2016
 #include <math.h>
 
 /* hypergeometric function 1F2 */
-long double h1f2(long double a, long double b, long double c, long double z);
+double h1f2(double a, double b, double c, double z);
 
 /* LegendreQ(m-1/2,n,cosh(x)) */
-long double LegendreQ(long double m, long double n, long double z);
+double LegendreQ(double m, double n, double z);
 
 /* derivative of LegendreQ(m-1/2,n,cosh(x)) */
-long double DLegendreQ(int m, int n, long double z);
+double DLegendreQ(int m, int n, double z);
 
 /* coordinate transformation r, z -> zeta */
-long double zetaf(long double rho, long double z, long double r0);
+double zetaf(double rho, double z, double r0);
 
 /* coordinate transformation r, z -> eta */
-long double etaf(long double rho, long double z, long double r0);
+double etaf(double rho, double z, double r0);
 
 int main () {
     int nmax, mmax, nphi, dim1, dim2;
@@ -44,24 +44,24 @@ int main () {
 //    fscanf (input, "%d", &nmax); // read n Maximum
     printf("MMax %d, NMax %d \n",mmax, nmax);
     dim1=(nmax+1)*(mmax+1); // coefficients dimensions 
-    long double cterm, rr;
-    long double *cc, *cs, *sc, *ss, *lsigma;
+    double cterm, rr;
+    double *cc, *cs, *sc, *ss, *lsigma;
     int **ccidx;
     int *midx;
     int *nidx;
 //    int ccidx[mmax+2][nmax+2];
-    long double zeta[30], eta[30];
-    long double zt, et, zt0, bzero, br, bz, bphi, bsize;
-    long double dataread[30], difrms, dzrms, ffrms, rmsd;
-    long double angles[30], radii[30], bfield[30][5];
+    double zeta[30], eta[30];
+    double zt, et, zt0, bzero, br, bz, bphi, bsize;
+    double dataread[30], difrms, dzrms, ffrms, rmsd;
+    double angles[30], radii[30], bfield[30][5];
     int i, j, id, idx, md, nd, mdx, ndx, prb, na, ma, qidx, vid;
-    long double datatmp, tp, tx, phi, wgt, iwgt, legq, dleq;
-    long double phidz, phidp, phide, phide1, phide2;
-    cc=(long double *)malloc(sizeof*cc*(dim1+2));
-    cs=(long double *)malloc(sizeof*cs*(dim1+2));
-    sc=(long double *)malloc(sizeof*sc*(dim1+2));
-    ss=(long double *)malloc(sizeof*ss*(dim1+2));
-    lsigma=(long double *)malloc(sizeof*lsigma*(nmax+2));
+    double datatmp, tp, tx, phi, wgt, iwgt, legq, dleq;
+    double phidz, phidp, phide, phide1, phide2;
+    cc=(double *)malloc(sizeof*cc*(dim1+2));
+    cs=(double *)malloc(sizeof*cs*(dim1+2));
+    sc=(double *)malloc(sizeof*sc*(dim1+2));
+    ss=(double *)malloc(sizeof*ss*(dim1+2));
+    lsigma=(double *)malloc(sizeof*lsigma*(nmax+2));
     ccidx=malloc((dim1+2)*sizeof(int*));
     if(ccidx==NULL){
         printf("out of memory\n");
@@ -90,7 +90,7 @@ int main () {
         radii[i]=45.;
     }
     for (i=1;i<=25;i++) {
-        printf("probe %d location r=%Lf, theta=%Lf \n",i,radii[i],angles[i]);
+        printf("probe %d location r=%g, theta=%g \n",i,radii[i],angles[i]);
     }
 
 //  calculate toroidal coordinates 
@@ -126,10 +126,10 @@ int main () {
     }
     for (idx=1;idx<=dim1*4;idx++) { // read coefficients
         fscanf (iname, "%d \t %d \t %d", &id, &ma, &na); 
-//        fscanf (input, "%d \t %d \t %d \t %Lf", &id, &ma, &na, &tp); 
-        fscanf (input, "%Lf", &tp); 
-        printf("%d, %d, %Lf \n", id, ccidx[ma][na], tp);
-//        if (tp!=tp) printf("%d, %d, %Lf \n", id, ccidx[ma][na], tp);
+//        fscanf (input, "%d \t %d \t %d \t %g", &id, &ma, &na, &tp); 
+        fscanf (input, "%g", &tp); 
+        printf("%d, %d, %g \n", id, ccidx[ma][na], tp);
+//        if (tp!=tp) printf("%d, %d, %g \n", id, ccidx[ma][na], tp);
         if (id==1) {
 //            printf("cc \n");
             cc[ccidx[ma][na]]=tp;
@@ -149,7 +149,7 @@ int main () {
         else {
             printf("error \n");
         }
-//        printf("%.20Lf \n",tp);
+//        printf("%.20g \n",tp);
     }
     fclose(input); // close data 
     fclose(iname); // close data 
@@ -157,7 +157,7 @@ int main () {
     nphi=0;
 //    datac=fopen("data40.txt", "r"); // open data for counting
     datac=fopen("data52.txt", "r"); // open data for counting
-    while(fscanf(datac, "%Lf", &datatmp)>0) {
+    while(fscanf(datac, "%g", &datatmp)>0) {
     nphi++;
     }
     nphi=nphi/26; // azimuthal angle + 25 probes 
@@ -166,14 +166,14 @@ int main () {
     dzrms=0.;
     ffrms=0.;
     for (id=1;id<=nphi;id++) { // loop over azimuthal slices 
-        fscanf (datafile, "%Lf", &dataread[0]); // read azimuthal angle in degrees 
+        fscanf (datafile, "%g", &dataread[0]); // read azimuthal angle in degrees 
         phi=dataread[0]*M_PI/180.;
         for (prb=1;prb<=25;prb++){ // loop over 25 probes 
-            fscanf (datafile, "%Lf", &dataread[prb]); 
+            fscanf (datafile, "%g", &dataread[prb]); 
             dataread[prb]=dataread[prb]*0.001+61.7400000;
             zt=zeta[prb];  // zeta coordinate at probe 
             et=eta[prb];  // eta coordinate at probe 
-//            printf("%Lf, %Lf \n", zt, et);
+//            printf("%g, %g \n", zt, et);
             wgt=sqrt(cosh(zt)-cos(et)); // weight func in toroidal coordinates 
             br=0.; bz=bzero; bphi=0.;
             for (i=1;i<=dim1;i++){
@@ -184,7 +184,7 @@ int main () {
                 legq=LegendreQ(md,nd,zt)/LegendreQ(md,nd,zt0);
 // Derivative of LegendreQ at probe
                 dleq=DLegendreQ(md,nd,zt)/LegendreQ(md,nd,zt0);
-//                printf("(%d, %d, %Lf) : %Lf, %Lf \n", md,nd,zt,legq, dleq);
+//                printf("(%d, %d, %g) : %g, %g \n", md,nd,zt,legq, dleq);
                 phidz=cc[ccidx[md][nd]]*cos(nd*phi)*cos(md*et)+  \
                       sc[ccidx[md][nd]]*cos(nd*phi)*sin(md*et)+  \
                       cs[ccidx[md][nd]]*sin(nd*phi)*cos(md*et)+  \
@@ -211,7 +211,7 @@ int main () {
                 bphi=bphi+(cosh(zt)-cos(et))/(rr*sinh(zt))*phidp;
             }
             bsize=sqrt(br*br+bz*bz+bphi*bphi);
-//            printf("%Lf \n",bsize);
+//            printf("%g \n",bsize);
             bfield[prb][4]=bsize;
             bfield[prb][1]=bz;
             bfield[prb][2]=br;
@@ -222,7 +222,7 @@ int main () {
             dzrms+=(dataread[prb]-bz)*(dataread[prb]-bz);
             ffrms+=(dataread[prb]-bzero)*(dataread[prb]-bzero);
         }
-        printf("Slice %d (phi=%Lf) of %d done (%Lf, %Lf, %Lf) \n",
+        printf("Slice %d (phi=%g) of %d done (%g, %g, %g) \n",
                 id, phi, nphi, difrms, dzrms, ffrms);
     }
     difrms=difrms/bzero/bzero/nphi/25;
@@ -231,15 +231,15 @@ int main () {
     difrms=sqrt(difrms)*1000000.;
     dzrms=sqrt(dzrms)*1000000.;
     ffrms=sqrt(ffrms)*1000000.;
-    printf("RMS fluctuation = %Lf ppm \n", ffrms);
-    printf("RMS difference (lin. approx.) = %Lf ppm \n", dzrms);
-    printf("RMS difference (real) = %Lf ppm \n", difrms);
+    printf("RMS fluctuation = %g ppm \n", ffrms);
+    printf("RMS difference (lin. approx.) = %g ppm \n", dzrms);
+    printf("RMS difference (real) = %g ppm \n", difrms);
     /* calculate coeffs */
 /*
     cterm=0.;
     printf ("\nComputing b-fields\n");
     for (id=1;id<=ndata;id++) { // loop over azimuthal slices 
-        fscanf (input, "%Lf", &datatmp); // read azimuthal angle in degrees 
+        fscanf (input, "%g", &datatmp); // read azimuthal angle in degrees 
         phi=datatmp*M_PI/180.;
         for (prb=1;prb<=25;prb++){ // loop over 25 probes 
             zt=zeta[prb];  // zeta coordinate at probe 
@@ -257,7 +257,7 @@ int main () {
             idx=0;
 // for each data end
         }
-    printf ("\nSlice %d (phi=%Lf) done\n", id, phi);
+    printf ("\nSlice %d (phi=%g) done\n", id, phi);
     }
 */
 /*
@@ -268,50 +268,50 @@ int main () {
 
     fprintf(out0, "NMax=%d;\n", nmax);
     fprintf(out0, "MMax=%d;\n", mmax);
-    fprintf(out0, "csqc0=%.20Lf;\n", cterm);
-    fprintf(out0, "csqc=(%.20Lf)+\n", cterm);
+    fprintf(out0, "csqc0=%.20g;\n", cterm);
+    fprintf(out0, "csqc=(%.20g)+\n", cterm);
 
     for (i=1;i<=dim1;i++) {
-        fprintf(out0, "(%.20Lf)*CC[%d,%d]+ \n", \
+        fprintf(out0, "(%.20g)*CC[%d,%d]+ \n", \
                 cc[i],ccidx[i][1],ccidx[i][2]);
-        if (ccidx[i][2]!=0) fprintf(out0, "(%.20Lf)*CS[%d,%d]+ \n", \
+        if (ccidx[i][2]!=0) fprintf(out0, "(%.20g)*CS[%d,%d]+ \n", \
                 cs[i],ccidx[i][1],ccidx[i][2]);
-        if (ccidx[i][1]!=0) fprintf(out0, "(%.20Lf)*SC[%d,%d]+ \n", \
+        if (ccidx[i][1]!=0) fprintf(out0, "(%.20g)*SC[%d,%d]+ \n", \
                 sc[i],ccidx[i][1],ccidx[i][2]);
         if (ccidx[i][1]!=0&&ccidx[i][2]!=0) fprintf(out0, \
-                "(%.20Lf)*SS[%d,%d]+ \n", \
+                "(%.20g)*SS[%d,%d]+ \n", \
                 ss[i],ccidx[i][1],ccidx[i][2]);
     }
     for (i=1;i<=dim2;i++) {
-        fprintf(out0, "(%.20Lf)*CC[%d,%d]*CC[%d,%d]+ \n",\
+        fprintf(out0, "(%.20g)*CC[%d,%d]*CC[%d,%d]+ \n",\
         cccc[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][4]!=0) fprintf(out0, \
-                    "(%.20Lf)*CC[%d,%d]*CS[%d,%d]+ \n",\
+                    "(%.20g)*CC[%d,%d]*CS[%d,%d]+ \n",\
         cccs[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][3]!=0) fprintf(out0, \
-                    "(%.20Lf)*CC[%d,%d]*SC[%d,%d]+ \n",\
+                    "(%.20g)*CC[%d,%d]*SC[%d,%d]+ \n",\
         ccsc[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][3]!=0&&ccccidx[i][4]!=0) fprintf(out0, \
-                    "(%.20Lf)*CC[%d,%d]*SS[%d,%d]+ \n",\
+                    "(%.20g)*CC[%d,%d]*SS[%d,%d]+ \n",\
         ccss[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][2]!=0&&ccccidx[i][4]!=0) fprintf(out0, \
-                    "(%.20Lf)*CS[%d,%d]*CS[%d,%d]+ \n",\
+                    "(%.20g)*CS[%d,%d]*CS[%d,%d]+ \n",\
         cscs[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][2]!=0&&ccccidx[i][3]!=0) fprintf(out0, \
-                    "(%.20Lf)*CS[%d,%d]*SC[%d,%d]+ \n",\
+                    "(%.20g)*CS[%d,%d]*SC[%d,%d]+ \n",\
         cssc[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][2]!=0&&ccccidx[i][3]!=0&&ccccidx[i][4]!=0) fprintf(out0, \
-                    "(%.20Lf)*CS[%d,%d]*SS[%d,%d]+ \n",\
+                    "(%.20g)*CS[%d,%d]*SS[%d,%d]+ \n",\
         csss[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][1]!=0&&ccccidx[i][3]!=0) fprintf(out0, \
-                    "(%.20Lf)*SC[%d,%d]*SC[%d,%d]+ \n",\
+                    "(%.20g)*SC[%d,%d]*SC[%d,%d]+ \n",\
         scsc[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][1]!=0&&ccccidx[i][3]!=0&&ccccidx[i][4]!=0) fprintf(out0, \
-                    "(%.20Lf)*SC[%d,%d]*SS[%d,%d]+ \n",\
+                    "(%.20g)*SC[%d,%d]*SS[%d,%d]+ \n",\
         scss[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][1]!=0&&ccccidx[i][2]&& \
             ccccidx[i][3]!=0&&ccccidx[i][4]!=0) fprintf(out0, \
-                    "(%.20Lf)*SS[%d,%d]*SS[%d,%d]",\
+                    "(%.20g)*SS[%d,%d]*SS[%d,%d]",\
         ssss[i], ccccidx[i][1],ccccidx[i][2],ccccidx[i][3],ccccidx[i][4]);
         if (ccccidx[i][1]!=0&&ccccidx[i][2]&& \
             ccccidx[i][3]!=0&&ccccidx[i][4]!=0&&i<dim2) fprintf(out0, "+ \n");
@@ -331,10 +331,10 @@ int main () {
 
 /* hypergeometric function 1F2 
    evaluated by truncating an infinite sum */
-long double h1f2(long double a, long double b, long double c, long double z){
+double h1f2(double a, double b, double c, double z){
 	int i, imax;
-	long double err, tol, si, sii, f1f2;
-	long double errabs, errrel;
+	double err, tol, si, sii, f1f2;
+	double errabs, errrel;
 	imax=100000000; // 10^8 maximum iterations
 	sii=1.; // initial term 
 	err=1.; // estimated uncertainty 
@@ -365,8 +365,8 @@ long double h1f2(long double a, long double b, long double c, long double z){
 /* Legendre function of the second kind, Q_{m+1/2}^n (cosh(z)).
    Normalized to remove gamma function.
    Regular inside the torus */
-long double LegendreQ(long double m, long double n, long double z){
-    long double lq;
+double LegendreQ(double m, double n, double z){
+    double lq;
 //    lq=pow(M_PI,.5)*exp(lgammal(m+n+.5))/(pow(2.,m+.5)*exp(lgammal(m+1.)));
     lq=pow(tanh(z),n)/pow(cosh(z),m+.5);
     lq=lq*h1f2(.5*(m+n+.5),.5*(m+n+1.5),m+1.,1./cosh(z)/cosh(z));
@@ -376,8 +376,8 @@ long double LegendreQ(long double m, long double n, long double z){
 /* Derivative of Q_{m+1/2}^n (cosh(z)).
    Normalized to remove gamma function.
  */
-long double DLegendreQ(int m, int n, long double z){
-    long double dlq, lq1, lq2;
+double DLegendreQ(int m, int n, double z){
+    double dlq, lq1, lq2;
     if (m==0) {
         dlq=-1/(8*pow(cosh(z),1.5))/sinh(z);
         dlq=dlq*pow(tanh(z),n);
@@ -397,15 +397,15 @@ long double DLegendreQ(int m, int n, long double z){
 }
 
 /* coordinate transformation r, z -> zeta */
-long double zetaf(long double rho, long double z, long double r0){
-    long double zetax;
+double zetaf(double rho, double z, double r0){
+    double zetax;
     zetax=atanhl(2.*rho*r0/(rho*rho+r0*r0+z*z));
     return zetax;
 }
 
 /* coordinate transformation r, z -> eta */
-long double etaf(long double rho, long double z, long double r0){
-    long double etax, xx;
+double etaf(double rho, double z, double r0){
+    double etax, xx;
     int i;
     xx=2.*r0*z/(rho*rho-r0*r0+z*z);
     if (fabsl(xx)<0.001) {
